requirements.txt
openai
python-dotenv
google-genai
json-repair
